app/pyproject.toml
[tool.poetry]
name = "SemSac"
version = "0.1.0"
description = ""
authors = ["AQUINO SANCHEZ Jorge <jorge.aquino@sparna.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
pyyaml = "^6.0.3"
rdflib = "^7.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
